Making migrations more upgrade friendly.

+++ src/database/migrations/2014_05_15_071630_create_module_media_memberships_table.php
@@ -12,17 +12,19 @@ class CreateModuleMediaMembershipsTable extends Migration
      */
     public function up()
     {
-        Schema::create('module_media_memberships', function(Blueprint $table) {
-            $table->increments('id');
-            $table->integer('module_id')->unsigned();
-            $table->foreign('module_id')->references('id')->on('modules')->onDelete('cascade');
-            $table->integer('media_id')->unsigned();
-            $table->foreign('media_id')->references('id')->on('medias')->onDelete('cascade');
-            $table->integer('membership_id')->unsigned();
-            $table->foreign('membership_id')->references('id')->on('memberships')->onDelete('cascade');
-            $table->unique(array('module_id', 'media_id', 'membership_id'));
-            $table->timestamps();
-        });
+        if (! Schema::hasTable('module_media_memberships')) {
+            Schema::create('module_media_memberships', function(Blueprint $table) {
+                $table->increments('id');
+                $table->integer('module_id')->unsigned();
+                $table->foreign('module_id')->references('id')->on('modules')->onDelete('cascade');
+                $table->integer('media_id')->unsigned();
+                $table->foreign('media_id')->references('id')->on('medias')->onDelete('cascade');
+                $table->integer('membership_id')->unsigned();
+                $table->foreign('membership_id')->references('id')->on('memberships')->onDelete('cascade');
+                $table->unique(array('module_id', 'media_id', 'membership_id'));
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -32,6 +34,6 @@ class CreateModuleMediaMembershipsTable extends Migration
      */
     public function down()
     {
-        Schema::drop('module_media_memberships');
+        Schema::dropIfExists('module_media_memberships');
     }
 }
