+++ src/database/migrations/2015_03_08_152556_create_coupon_product_table.php
@@ -12,14 +12,16 @@ class CreateCouponProductTable extends Migration {
 	 */
 	public function up()
 	{
-		Schema::create('coupon_product', function(Blueprint $table)
-		{
-			$table->increments('id');
-            $table->integer('coupon_id')->unsigned();
-            $table->integer('product_id')->unsigned();
-            $table->foreign('coupon_id')->references('id')->on('coupons')->onDelete('cascade');
-            $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
-		});
+        if (! Schema::hasTable('coupon_product')) {
+            Schema::create('coupon_product', function(Blueprint $table)
+            {
+                $table->increments('id');
+                $table->integer('coupon_id')->unsigned();
+                $table->integer('product_id')->unsigned();
+                $table->foreign('coupon_id')->references('id')->on('coupons')->onDelete('cascade');
+                $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
+            });
+        }
 	}
 
 	/**
@@ -29,7 +31,7 @@ class CreateCouponProductTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('coupon_product');
+		Schema::dropIfExists('coupon_product');
 	}
 
 }
